Add documentation to schema + minor refactor
